Replace the PathFuncs interface magic with a plain map
The strange interface magic done in configurers to allow overriding the
path functions is something I never quite understood how it actually
works and why it is necessary. Here the mechanism is replaced with a
simple map[string]string and getter functions + a setter because in the
dry-run mechanism it was necessary to programmatically override the
K0sBinaryPath() at run time.

Signed-off-by: Kimmo Lehto <klehto@mirantis.com>